Inline the submodules of `cli_rework`
